requirements.txt
Django>=4.2,<5.1

djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.5
django-storages==1.14.2
django-redis==5.4.0
django-ratelimit==4.1.0
django-modeltranslation==0.18.11
django-rosetta==0.10.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0
redis==5.0.1

# Authentication & Security
djangorestframework-simplejwt==5.4.0
django-guardian>=2.4,<3
cryptography>=41,<47
pyotp>=2.9,<3
qrcode>=7.4,<8
stripe>=7.0,<9
argon2-cffi>=23.1.0
# File handling
Pillow==10.1.0
python-magic==0.4.27

# Email & SMS
celery==5.3.4
django-celery-beat>=2.6,<3
twilio>=8.12,<9

# AWS/Cloud Storage
boto3==1.34.0
botocore==1.34.0

# API Documentation
drf-spectacular==0.27.0

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# Code Quality
black==23.12.0
ruff==0.1.8
pre-commit==3.6.0

# Monitoring & Logging
sentry-sdk==1.39.2
structlog==23.2.0
prometheus-client==0.19.0
django-prometheus==2.3.1

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-decouple==3.8
python-slugify==8.0.1
phonenumbers==8.13.26
django-phonenumber-field>=7.0,<9
babel==2.14.0